Make member variables volatile
Make non-final members of FixedIntervalRefillStrategy and
TokenBucketImpl volatile, to ensure that changes to their values
will be immediately visible in other threads. Since all updates to
those member variables occur in synchronized methods, this change
is sufficient to prevent race conditions.

Tweak javadoc for TokenBucket.RefillStrategy.refill() to make it
clearer that the method should not return a fixed value.